批量处理不应该覆盖已处理的预警

+++ dataanalysis/src/main/java/com/aizhixin/cloud/dataanalysis/alertinformation/service/AlarmHandlingService.java
@@ -114,18 +114,19 @@ public class AlarmHandlingService {
             for (Map map : idList) {
                 if (null != map && null != map.get("ID")) {
                     WarningInformation warningInformation = alertWarningInformationService.getOneById(map.get("ID").toString());
-                    Map<String, String> maps = domain.getDealTypes();
-                    if (null != maps) {
-
-                        for (Map.Entry<String, String> e : maps.entrySet()) {
-                            operaionRecordService.deleteByWarningInformationIdAndDealType(map.get("ID").toString(),Integer.parseInt(e.getKey()));
-                            OperationRecord operationRecord1 = new OperationRecord();
-                            operationRecord1.setWarningInformationId(map.get("ID").toString());
-                            operationRecord1.setOrgId(warningInformation.getOrgId());
-                            operationRecord1.setOperationTime(new Date());
-                            operationRecord1.setDealType(Integer.parseInt(e.getKey()));
-                            operationRecord1.setProposal(e.getValue());
-                            operaionRecordService.save(operationRecord1);
+                    if (warningInformation.getWarningState() != 20) {
+                        Map<String, String> maps = domain.getDealTypes();
+                        if (null != maps) {
+                            for (Map.Entry<String, String> e : maps.entrySet()) {
+                                operaionRecordService.deleteByWarningInformationIdAndDealType(map.get("ID").toString(), Integer.parseInt(e.getKey()));
+                                OperationRecord operationRecord1 = new OperationRecord();
+                                operationRecord1.setWarningInformationId(map.get("ID").toString());
+                                operationRecord1.setOrgId(warningInformation.getOrgId());
+                                operationRecord1.setOperationTime(new Date());
+                                operationRecord1.setDealType(Integer.parseInt(e.getKey()));
+                                operationRecord1.setProposal(e.getValue());
+                                operaionRecordService.save(operationRecord1);
+                            }
                         }
                     }
                 }
@@ -150,23 +151,23 @@ public class AlarmHandlingService {
             String[] idArray = batchDealDomain.getWarningInformationIds();
             for (String id : idArray) {
                 WarningInformation warningInformation = alertWarningInformationService.getOneById(id);
-                warningInformation.setLastModifiedDate(new Date());
-                warningInformation.setWarningState(AlertTypeConstant.ALERT_PROCESSED);
-                alertWarningInformationService.save(warningInformation);
-                Map<String, String> map = batchDealDomain.getDealTypes();
-                if (null != map) {
-
-                    for (Map.Entry<String, String> e : map.entrySet()) {
-                        //删除历史的处理意见
-                        operaionRecordService.deleteByWarningInformationIdAndDealType(id,Integer.parseInt(e.getKey()));
-
-                        OperationRecord operationRecord1 = new OperationRecord();
-                        operationRecord1.setWarningInformationId(id);
-                        operationRecord1.setOrgId(warningInformation.getOrgId());
-                        operationRecord1.setOperationTime(new Date());
-                        operationRecord1.setDealType(Integer.parseInt(e.getKey()));
-                        operationRecord1.setProposal(e.getValue());
-                        operaionRecordService.save(operationRecord1);
+                if (warningInformation.getWarningState() != 20) {
+                    warningInformation.setLastModifiedDate(new Date());
+                    warningInformation.setWarningState(AlertTypeConstant.ALERT_PROCESSED);
+                    alertWarningInformationService.save(warningInformation);
+                    Map<String, String> map = batchDealDomain.getDealTypes();
+                    if (null != map) {
+                        for (Map.Entry<String, String> e : map.entrySet()) {
+                            //删除历史的处理意见
+                            operaionRecordService.deleteByWarningInformationIdAndDealType(id, Integer.parseInt(e.getKey()));
+                            OperationRecord operationRecord1 = new OperationRecord();
+                            operationRecord1.setWarningInformationId(id);
+                            operationRecord1.setOrgId(warningInformation.getOrgId());
+                            operationRecord1.setOperationTime(new Date());
+                            operationRecord1.setDealType(Integer.parseInt(e.getKey()));
+                            operationRecord1.setProposal(e.getValue());
+                            operaionRecordService.save(operationRecord1);
+                        }
                     }
                 }
             }
